requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx>=0.25.0
motor>=3.3.0
pymongo>=4.9,<5.0
dnspython>=2.6.0

# Python 3.11+ required
